Move 3 Ast classes into seperate files
