+++ app/lib/celery_script/ast_base.rb
@@ -1,0 +1,9 @@
+module CeleryScript
+  # Abstract class that AstLeaf and AstNode inherit from.
+  # Use this for shared leaf/node behavior.
+  class AstBase
+    def invalidate!(message = "Unspecified type check error.")
+      raise CeleryScript::TypeCheckError, message
+    end
+  end
+end
