+++ app/lib/celery_script/ast_leaf.rb
@@ -1,0 +1,8 @@
+module CeleryScript
+  class AstLeaf < AstBase
+    attr_reader :kind, :value, :parent
+    def initialize(parent, value, kind)
+      @parent, @value, @kind = parent, value, kind
+    end
+  end
+end
